requirements.txt
# 游戏自动化测试框架依赖包
airtest>=1.2.10
pytest>=7.0.0
pytest-allure>=1.0.0
allure-pytest>=2.12.0
allure-python-commons>=2.12.0